edstem-scheduler/requirements.txt
boto3==1.18.11
botocore==1.21.11
certifi==2021.5.30
charset-normalizer==2.0.4
click==7.1.2
colorama==0.4.4
idna==3.2
importlib-metadata==4.6.3
itsdangerous==1.1.0
jmespath==0.10.0
MarkupSafe==2.0.1
python-dateutil==2.8.2
python-dotenv==0.19.0
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
typing-extensions==3.10.0.0
urllib3==1.26.6
zipp==3.5.0
